refactor: generate shared types module instead of duplicating (#523)
