+++ sdk/src/index.ts
@@ -10,13 +10,8 @@ import {
   scValToNative,
 } from "@stellar/stellar-sdk";
 
-// === Types
-
-export interface NetworkConfig {
-  rpcUrl: string;
-  networkPassphrase: string;
-  contractId: string;
-}
+export * from "./types";
+import { NetworkConfig, Bounty, BountyMeta, Contributor, CreateBountyParams } from "./types";
 
 export const TESTNET: Omit<NetworkConfig, "contractId"> = {
   rpcUrl: "https://soroban-testnet.stellar.org",
@@ -27,37 +22,6 @@ export const MAINNET: Omit<NetworkConfig, "contractId"> = {
   rpcUrl: "https://mainnet.stellar.validationcloud.io/v1/XCa...",
   networkPassphrase: Networks.PUBLIC,
 };
-
-export interface Bounty {
-  creator: string;
-  rewardAmount: bigint;
-  rewardToken: string;
-  assignees: Array<{ address: string; shareBp: number }>;
-  maxAssignees: number;
-  status: string;
-  minReputation: number;
-  deadline: number | null;
-}
-
-export interface BountyMeta {
-  title: string;
-  description: string;
-}
-
-export interface Contributor {
-  address: string;
-  reputation: number;
-  totalEarned: bigint;
-  contributionCount: number;
-  activeClaims: number;
-  metadata: string | null;
-}
-
-export interface CreateBountyParams {
-  creator: string;
-  title: string;
-  description: string;
-  rewardAmount: bigint;
   rewardToken: string;
   minReputation: number;
   deadline: number | null;
